+++ BatteryInterface.h
@@ -6,8 +6,8 @@ class BatteryInterface
 public:
 	virtual int stepsLeft() const = 0;
 	virtual bool isBatteryEmpty() const = 0;
-	virtual uint32_t setBatteryLevelInSteps(int stepsToSet) = 0;
-	virtual uint32_t decrementBatterySingleStep() = 0;
-	virtual uint32_t chargeBatteryDuringSingleStep() = 0;
+	virtual float setBatteryLevelInSteps(float stepsToSet) = 0;
+	virtual float decrementBatterySingleStep() = 0;
+	virtual float chargeBatteryDuringSingleStep() = 0;
 };
